Add in event details and list views, registration page, filled out delegate event views
