+++ tools/templates/tools/approve-delegated-event/approve.html
@@ -1,0 +1,25 @@
+{% extends "base.html" %} {% block page_content %}
+
+{% include "delegated-events/details.html" %}
+<h1 class="text-xl py-8">Approve Delegated Event</h1>
+<form
+  action=""
+  method="post"
+>
+  {% csrf_token %} {{ form.non_field_errors }} {% for field in form %}
+  <div>
+    <div class="py-2">
+      <div>{{ field.label_tag }}</div>
+      <div>{{ field }}</div>
+    </div>
+    <div class="text-danger">{{ field.errors }}</div>
+  </div>
+  {% endfor %}
+<input
+    type="submit"
+    name="submit-new-event"
+    value="Submit"
+    class="bg-primary text-white p-2 my-3"
+  />
+</form>
+{% endblock page_content %}
