+++ tools/templates/tools/approve-delegated-event/unresolveable.html
@@ -1,0 +1,22 @@
+{% extends "base.html" %} {% block page_content %}
+<p> Your event request could not be created due to the following unresolveable conflicts. </p>
+<hr>
+<p>
+    <ul>
+    {% for conflict in conflicts %}
+    <li>
+        {% if conflict.type == 0 %}
+            Zoom Conflict 
+            {%if conflict.zoomUser %} 
+                ({{conflict.zoomUser}})
+            {% endif %}
+            :
+        {% else %}
+            Google Calendar Conflict: 
+        {% endif %}
+        {{ conflict.title }} - {{ conflict.start }} - {{conflict.end}}
+    </li>
+    {% endfor %}
+    </ul>
+</p>
+{% endblock page_content %}
