+++ tools/templates/tools/events/details.html
@@ -21,5 +21,10 @@
     Description: {{ object.description }} <br>
     Instructions: {{ object.instructions }} <br>
     <br>
+    AN Manage Link: {{ object.anManageLink }} <br>
+    An Share Link : {{ object.anShareLink }} <br>
+    GCal Link : {{ object.gCalLink }} <br>
+    Zoom Link ({{ object.zoomAccount }}) : {{ object.zoomLink }} <br>
+
 </p>
 {% endblock page_content %}
