+++ tools/templates/tools/new-delegated-event/new.html
@@ -11,7 +11,7 @@
     </select>
     <input type="submit" value="Set">
 </form> {% endcomment %}
-<h1 class="text-xl py-8">Create an Event</h1>
+<h1 class="text-xl py-8">Create a Delegated Event</h1>
 <form
   action=""
   method="post"
